requirements.txt
PyQt6>=6.4.0
pyinstaller>=6.0.0
requests>=2.28.0
pandas>=1.5.0
gpxpy>=1.5.0
polyline>=2.0.0
keyring>=23.0.0
pytest>=7.0.0
pytest-qt>=4.2.0
pytest-cov>=4.0.0
Pillow>=9.0.0
pydantic>=2.0
cachetools>=5.0
pydantic-settings>=2.0
python-dotenv>=1.0